FlugTicketSystem perfektioniert. Vollständig sinnvolle Quellcode kommentierung nur Ausgaben in der Boundaryklasse. Funktionen und attribute exakt wie im UML, außer operatoren, getter, setter und Konstruktoren. Man kann sich einloggen, registrieren, ein flug per nummer suchen, einen flug per flughäfen suchen. Ticketskaufen und belegt damit tasächlich kapazitäten in dem flugzeug. manuell fluege aus der datei laden.

+++ FlugTicketSystem/Ticket.h
@@ -37,12 +37,13 @@ public:
     Reiseklasse getKlasse() {
         return m_klasse;
     };
+    
 
     inline bool operator>(const Ticket& rigth) const;
     inline bool operator<(const Ticket& rigth) const;
 private:
 
-    unsigned int m_preis;
+    float m_preis;
     int m_ticketnummer;
     Reiseklasse m_klasse;
     static int counter;
